Added history

+++ src/main/java/pl/adamklimko/kkp/repository/CleanedRoomsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import pl.adamklimko.kkp.model.CleanedRooms;
 
 @Repository
-public interface CleanedRoomsRepository extends CrudRepository<CleanedRooms, Integer> {
+public interface CleanedRoomsRepository extends CrudRepository<CleanedRooms, Long> {
 }
