+++ src/main/java/pl/adamklimko/kkp/repository/ProfileRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import pl.adamklimko.kkp.model.user.Profile;
 
 @Repository
-public interface ProfileRepository extends CrudRepository<Profile, Integer>{
+public interface ProfileRepository extends CrudRepository<Profile, Long>{
 }
